pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pcbko"
version = "0.1"
description = "Open source PCB design assistant with AI-driven KiCad output generation."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["pcb", "kicad", "eda", "ai", "pcb design"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]
dependencies = [
  "openai>=1.40.0",
  "pydantic>=2.8.2",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "skidl>=1.1.0",
  "matplotlib>=3.8.0",
  "Pillow>=10.0.0",
  "PySide6>=6.10.0"
]
